refactor: Refatora metodo para exibir as estatisticas, usando stream com mapToDouble para mudar os dados de BigDecimal para double, finaliza a stream com summaryStatiscs para obter as estatisticas e passar no DTO refatorado.

+++ src/main/java/dev/eu/desafioitau/dto/response/EstatisticaResponse.java
@@ -7,9 +7,9 @@ import java.math.BigDecimal;
 @Builder
 public record EstatisticaResponse(
         Long count,
-        BigDecimal sum,
-        BigDecimal avg,
-        BigDecimal max,
-        BigDecimal min
+        double sum,
+        double avg,
+        double max,
+        double min
 ) {
 }
